finished implementation
