+++ __mocks__/axios.js
@@ -1,0 +1,5 @@
+const get = url => {
+  return Promise.resolve({ data: { title: 'delectus aut autem' } });
+};
+
+exports.get = get;
