+++ __mocks__/http.js
@@ -1,0 +1,5 @@
+const fetchData = () => {
+  return Promise.resolve({ title: 'delectus aut autem' });
+};
+
+exports.fetchData = fetchData;
