+++ util.test.js
@@ -1,0 +1,9 @@
+// jest.mock('./http');
+
+const { loadTitle } = require('./util');
+
+test('should print an uppercase text', () => {
+  loadTitle().then(title => {
+    expect(title).toBe('DELECTUS AUT AUTEM');
+  });
+});
